refactor(codebuild-source-credential): convert to libnuke resource format
